% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAdministrationFuncs.R
\name{adm.savePkgLib}
\alias{adm.savePkgLib}
\title{Save existing package library to a temporary file.}
\usage{
adm.savePkgLib()
}
\value{
Nothing; just a message indication completion
}
\description{
Save existing package library to a temporary file.
}
\details{
When upgrading your R version, run this first
   to save your existing package library. It can be
   restored after the upgrade.
}
