% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statFuncs.R
\name{betaDiffZ}
\alias{betaDiffZ}
\title{Z score and 2-tailed p of the difference between two regression betas}
\usage{
betaDiffZ(b1, b2, seb1, seb2)
}
\arguments{
\item{b1}{A regression coefficient}

\item{b2}{A regression coefficient}

\item{seb1}{The standard error of b1}

\item{seb2}{The standard error of b2}
}
\value{
Prints Z and 2-tailed p value to the console
}
\description{
Z score and 2-tailed p of the difference between two regression betas
}
