% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAdministrationFuncs.R
\name{changePkgLibLocation}
\alias{changePkgLibLocation}
\title{Changes library location for package installation}
\usage{
changePkgLibLocation()
}
\value{
A message with the new library location to be used for installation
}
\description{
Changes library location for package installation
}
\details{
Wrapper for the R function .libPaths
}
