% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modInfo.R
\name{EfxList}
\alias{EfxList}
\title{Creates a table of the terms in a regression output object (of class lme or
glmmTMB)}
\usage{
EfxList(resultObj)
}
\arguments{
\item{resultObj}{A regression output object of class lme or glmmTMB}
}
\value{
a tbl_df with two columns: the term (effect) number, and the term
name (description)
}
\description{
Creates a table of the terms in a regression output object (of class lme or
glmmTMB)
}
\note{
This function is particularly useful when doing moderation analysis
(e.g. with the function JNSiena5.3way), as the latter refers to terms in
the model by their number (in order, as used for vectors of lists of
coefficients, rows and cols of the vcov matrix, etc.)
}
