% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleFuncs.R
\name{exampleStruc0Code}
\alias{exampleStruc0Code}
\title{Make a 'Matrix' class matrix, use row/col names to assign values}
\usage{
exampleStruc0Code()
}
\value{
Nothing. This is just example code
}
\description{
Make a 'Matrix' class matrix, use row/col names to assign values
}
\details{
This example assigns RSiena 'structural zeros' to
   edges where individuals have different houses. The house
   information is encoded in a set (list, perhaps) of vectors
   that show which rows (& cols) belong in which houses.
}
