% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFuncs.R
\name{find_var}
\alias{find_var}
\title{Finds a variable in a data.frame by matching a pattern, etc.}
\usage{
find_var(
  data,
  pattern,
  ignore.case = TRUE,
  search = c("name_label", "name_value", "label_value", "name", "label", "value", "all"),
  out = c("table", "df", "index"),
  fuzzy = FALSE,
  regex = FALSE
)
}
\arguments{
\item{data}{A dataframe with the variables of interest}

\item{pattern}{A string with the pattern to look for, or a regular expression}

\item{ignore.case}{Logical, TRUE => match irrespective of case.}

\item{search}{A string indicating what to look for. Can be name_label,
name_value, label_value, name, or label. Default is name_label}

\item{out}{A string requesting type of output; may be a table ("table"),
dataframe ("df"), or index ("index").}

\item{fuzzy}{Logical, if TRUE use fuzzy grep.}

\item{regex}{Logical, if TRUE pattern is a regex.}
}
\value{
A table, dataframe or index (see @out).
}
\description{
This functions finds variables in a data frame, which variable
 names or variable (and value) label attribute match a specific pattern.
 Regular expression for the pattern is supported. Copied from package
 >sjmisc<, which has a lot of other useful functions of this general
 type.
}
\details{
This function searches for pattern in data's column names and
 - for labelled data - in all variable and value labels of data's variables
 (see sjmisc::get_label for details on variable labels and labelled data).
 Regular expressions are supported as well, by simply using
 pattern = stringr::regex(...) or regex = TRUE.
}
