% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoringFuncs.R
\name{fixPSQNA}
\alias{fixPSQNA}
\title{Internal function used by PSQIScore}
\usage{
fixPSQNA(inTb)
}
\arguments{
\item{inTb}{A tbl_df with SSID, then all raw PSQI numeric items in order.\cr
Hence it is N x 19 (SSID + 18 PSQI items).\cr
inTb cols 2 & 4 must be of type c('hms', 'difftime' )\cr
inTb cols 3 and 5:19 must include either 'double' or 'numeric'
   in their class.\cr
intTb col 1 (the ID column), the PSQI ID is normally some numeric
but can be anything.\cr
The 18 input items (ignoring SSID, the first column) must be
included in this order, though they can be named anything:\cr
- psq1\cr
- psq2\cr
- psq3\cr
- psq4\cr
- psq5a to psq5j (10 items)\cr
- psq6\cr
- psq7\cr
- psq8\cr
- psq9\cr}
}
\value{
A tbl_df with the same format as the input tbl_df, only with NAs
     re-scored to 0's, as appropriate.
}
\description{
Internal function used by PSQIScore
}
