% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAdministrationFuncs.R
\name{folderLocation}
\alias{folderLocation}
\title{Determines the folder loc on this machine for reading/writing}
\usage{
folderLocation()
}
\value{
A string describing a folder location, specific to the
   current machine.
}
\description{
Determines the folder loc on this machine for reading/writing
}
\details{
This function is used to determine where the R package
   backup should be written, and where to retrieve it from.
}
