% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statFuncs.R
\name{GeodesicDistribution}
\alias{GeodesicDistribution}
\title{Calculates geodesic distribution stats for sienaGOF}
\usage{
GeodesicDistribution(i, data, sims, period, groupName, varName, levls = c(1:5,
  Inf), cumulative = TRUE, ...)
}
\description{
For details see ?sna::geodist
}
\note{
The default for \code{levls} reflects that geodesic distances
larger than 5 do not differ appreciably with respect to interpretation.

Levels of the result are named, and used in the \code{plot} method
}
