% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFuncs.R
\name{getglmmTMBFit}
\alias{getglmmTMBFit}
\title{Pulls glmmTMB model results with t and p-vals}
\usage{
getglmmTMBFit(fitObj, pVal = TRUE, tails = 2)
}
\arguments{
\item{fitObj}{A 'glmmTMB' object (output from glmmTMB)}

\item{pVal}{Logical: if TRUE (default), calculates the t-ratio and
normal-distribution p-value (1 or 2 tailed depending on the 'tails'
parameter) for each effect.}

\item{tails}{Numeric: default is 2; enter 1 if 1-tailed pval preferred.}
}
\value{
A tibble with the 5 columns described earlier (fx name, beta, SE,
  tstat, pval).
}
\description{
Extracts effect names, betas (estimated parameters) and
  std errors from a glmmTMB results object, calculates t-statistics and one
  or two-tailed p-values for each effect, and exports these into a tbl
}
\details{
Uses a version of JMLUtils::pValT (pvT) which does not output
  anything, just calculates 1 or 2-tailed p-vals.
}
