% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFuncs.R
\name{getRSFit}
\alias{getRSFit}
\title{Pulls RSiena model results with t and p-vals}
\usage{
getRSFit(fitObj, pVal = TRUE, tails = 2)
}
\arguments{
\item{fitObj}{An RSiena 'sienaFit' object (output from siena07)}

\item{pVal}{Logical: if TRUE (default), calculates the t-ratio and
normal-distribution p-value (1 or 2 tailed depending on the 'tails'
parameter) for each effect.}

\item{tails}{Numeric: default is 2; enter 1 if 1-tailed pval preferred.}
}
\value{
A tibble with the 7 columns described earlier.
}
\description{
Extracts effect numbers, effect types, effect names,
  thetas (estimated parameters) and standard errors from a
  sientFit object, then calculates t-statistics and one or two-tailed
  p-values for each effect.
}
\details{
Uses a version of JMLUtils::pValT (pvT) which does not output
  anything, just calculates 1 or 2-tailed p-vals.
}
