% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFuncs.R
\name{inputSPSS}
\alias{inputSPSS}
\title{Read an SPSS-format network file into a 'wide-format' data.table}
\usage{
inputSPSS(pInput = "")
}
\arguments{
\item{pInput}{A character string giving the full path to the SPSS file
to be read in. If blank, the user is asked to select an input file
from the file system.}
}
\value{
A tibble in 'wide' format ().
   Wide Format' means that each chooser has only 1 row per wave, and the
   relationship choices directed to other house members are all contained
   in that one row. The data.table is basically an image of the SPSS file
   (.sav) that it reads.
}
\description{
This is usually the first function run when creating a network analysis.
}
