% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFuncs.R
\name{install.glmmADMB}
\alias{install.glmmADMB}
\title{Installs glmmADMB from source}
\usage{
install.glmmADMB()
}
\value{
Nothing. Simply installs the package in the default package lib.
}
\description{
Installs glmmADMB package from source. This is needed
  whenever a new version of R is released that requires an update
  to the package library (viz. the "X" or "y" in version X.y.z).
  (You cannot simply transfer the package using the adm.savePkgLib
  and adm.restorePkgLib functions from this package)
}
\details{
Be sure you have the new version of R installed and are using
  it (if this is relevant) before running.
}
