% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statFuncs.R
\name{JNSiena.2way}
\alias{JNSiena.2way}
\title{Calculates Johnson-Neyman significance regions and MOM signficance
  values (raw and Bonferroni-Holm adjusted) for SAOM 2-way
  interactions.}
\usage{
JNSiena.2way(
  siena07out,
  theta1,
  theta2,
  thetaInt,
  theta1vals,
  theta2vals,
  sigRegion = TRUE,
  alpha = 0.05
)
}
\arguments{
\item{siena07out}{An object of class 'sienaFit'}

\item{theta1}{Number (from RSiena output) of 1st effect involved in
a 2-way interaction effect included in 'siena07out'. This effect
will be the primary predictor for the first table of pick-a-point
results in the output, and the moderator in the second such table.}

\item{theta2}{Number (from RSiena output) of 2st effect involved in
a 2-way interaction effect included in 'siena07out'. This effect
will be the moderator for the first table of pick-a-point
results in the output, and the primary predictor in the second
such table.}

\item{thetaInt}{Number (from RSiena output) of the multiplicative
interaction of the two effects identified by theta1 and theta2.}

\item{theta1vals}{A vector of numerical values of effect 1 for which raw and
B-H-adjusted significance should be obtained, applying both to
effect 1 as a predictor (in the 1st output table) and as a
moderator (in the 2nd output table).}

\item{theta2vals}{A vector of numerical values of effect 2 for which raw and
B-H-adjusted significance should be obtained, applying both to
effect 2 as a predictor (in the 2nd output table) and as a
moderator (in the 1st output table).}

\item{sigRegion}{Logical; default TRUE. If TRUE, Johnson-Neyman (raw)
significance limits are calculated for the case of effect 1 as
primary and effect 2 as moderator, and vice versa.}

\item{alpha}{Number (default = 0.05) corresponding the 2-tailed significance
level used to calculate whether a moderated effect is significant for a
given value (for points in theta1vals and theta2vals) or range
(for JN significance region computation).}
}
\value{
A list of 5 elements.
    \enumerate{
       \item{1. *thetas* Three matrices of the predicted theta values
       of each of predictors 1, 2, and 3, calculated on each pair of
       values in the range (or discrete values) provided for each
       of the remaining two moderators.}
       \item{2. *standard_errors* SEs calculated conditionally in the
       same way as for the thetas.}
       \item{3. *p_values* P-values calculated conditionally in the
       same way as for the thetas.}
       \item{4. *significance* TRUE if the p-value falls within the
       Region Of Significance, as specified by alpha and, possibly also
       any adjustment for multiple tests; otherwise FALSE. Calculated
       in a matrix with the same conditions as for thetas, etc.}
       \item{5. *plots* A list with three class gg and ggplot elements as
       its members, corresponding to each of the three predictors being
       taken as the primary predictor, with the same ordering as for the
       previous list elements (theta, etc.). }
       }
}
\description{
Calculates Johnson-Neyman significance regions and MOM signficance
  values (raw and Bonferroni-Holm adjusted) for SAOM 2-way
  interactions.
}
