% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statFuncs.R
\name{JNSiena.3way}
\alias{JNSiena.3way}
\title{Calculates Johnson-Neyman significance regions and MOM signficance
  values (raw and Bonferroni-Holm adjusted) for SAOM 2-way
  interactions.}
\usage{
JNSiena.3way(
  siena07out,
  theta1,
  theta2,
  theta3 = NULL,
  thetaInt12,
  thetaInt13 = NULL,
  thetaInt23 = NULL,
  thetaInt123 = NULL,
  theta1Vals,
  theta2Vals,
  theta3Vals = NULL,
  control_fdr = FALSE,
  alpha = 0.05,
  round_res = 3,
  color_low = "white",
  color_high = "#000066",
  color_values = "limegreen"
)
}
\arguments{
\item{siena07out}{An object of class 'sienaFit'}

\item{theta1}{Number (from RSiena output) of 1st effect involved in
a 2/3-way interaction effect included in 'siena07out'.}

\item{theta2}{Number (from RSiena output) of 2st effect involved in
a 2/3-way interaction effect included in 'siena07out'.}

\item{theta3}{Number (from RSiena output) of 3rd effect involved in
a 2/3-way interaction effect included in 'siena07out'. If it is
not included in the function call, a 2-way interaction is
assumed involving theta1 and theta2}

\item{thetaInt12}{Number (from RSiena output) of the multiplicative
interaction of the two effects identified by theta1 and theta2.
This is always required.}

\item{thetaInt13}{If theta3 is included, the number (from RSiena output)
of the multiplicative interaction of the effects identified as
theta1 and theta3.}

\item{thetaInt23}{If theta3 is included, the number (from RSiena output)
of the multiplicative interaction of the effects identified as
theta2 and theta3.}

\item{thetaInt123}{If theta3 is included, the number (from RSiena output)
of the multiplicative interaction of the effects identified as
theta1, theta2 and theta3.}

\item{alpha}{Number (default = 0.05) corresponding the 2-tailed significance
level used to calculate whether a moderated effect is significant for a
given value (for points in theta1vals, theta2vals, and theta3vals)
ranges.}

\item{theta1vals}{A vector of numerical values of effect 1 for which raw and
B-H-adjusted significance should be obtained, applying both to
effect 1 as a predictor (in the 1st output table) and as a
moderator (in the 2nd output table).}

\item{theta2vals}{A vector of numerical values of effect 2 for which raw and
B-H-adjusted significance should be obtained, applying both to
effect 2 as a predictor (in the 2nd output table) and as a
moderator (in the 1st output table).}

\item{theta3vals}{A vector of numerical values of effect 3 for which
significance should be obtained, applying both to
 effect 3 as a predictor and as a moderator.}
}
\value{
A list of 5 main elements: Theta, standard_errors, p_values,
    significance, and plots. Each of these elements will have 2 (if 2-way
    analysis) or 3 (if 3-way analysis) sub-elements giving the
    moderator-conditional predicted values of theta, SE, p, and
    significance as tables (with each variable having a turn at being
    the focal predictor, with the other(s) as moderator(s). The plots
    element contains (obviously) heat map graphs showing these
    results graphically.
}
\description{
Calculates Johnson-Neyman significance regions and MOM signficance
  values (raw and Bonferroni-Holm adjusted) for SAOM 2-way
  interactions.
}
