% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statFuncs.R
\name{JNSiena}
\alias{JNSiena}
\title{Calculates Johnson-Neyman significance regions and MOM signficance
  values (raw and Bonferroni-Holm adjusted) for SAOM 2-way
  interactions.}
\usage{
JNSiena(
  siena07out,
  theta1,
  theta2,
  thetaInt,
  theta1vals,
  theta2vals,
  sigRegion = TRUE,
  alpha = 0.05
)
}
\arguments{
\item{siena07out}{An object of class 'sienaFit'}

\item{theta1}{Number (from RSiena output) of 1st effect involved in
a 2-way interaction effect included in 'siena07out'. This effect
will be the primary predictor for the first table of pick-a-point
results in the output, and the moderator in the second such table.}

\item{theta2}{Number (from RSiena output) of 2st effect involved in
a 2-way interaction effect included in 'siena07out'. This effect
will be the moderator for the first table of pick-a-point
results in the output, and the primary predictor in the second
such table.}

\item{thetaInt}{Number (from RSiena output) of the multiplicative
interaction of the two effects identified by theta1 and theta2.}

\item{theta1vals}{A vector of numerical values of effect 1 for which raw and
B-H-adjusted significance should be obtained, applying both to
effect 1 as a predictor (in the 1st output table) and as a
moderator (in the 2nd output table).}

\item{theta2vals}{A vector of numerical values of effect 2 for which raw and
B-H-adjusted significance should be obtained, applying both to
effect 2 as a predictor (in the 2nd output table) and as a
moderator (in the 1st output table).}

\item{sigRegion}{Logical; default TRUE. If TRUE, Johnson-Neyman (raw)
significance limits are calculated for the case of effect 1 as
primary and effect 2 as moderator, and vice versa.}

\item{alpha}{Number (default = 0.05) corresponding the 2-tailed significance
level used to calculate whether a moderated effect is significant for a
given value (for points in theta1vals and theta2vals) or range
(for JN significance region computation).}
}
\value{
A list. The first two elements of the list are dataframes containing
    both raw and B-H-adjusted significance (the former a p-value, the latter
    simply TRUE or FALSE) of the primary effect for each value of the
    moderator given by thetaXvals (X=1 or 2). The first such table treats
    effect 1 as primary, moderated by effect 2, and the second table is the
    reverse. Additionally, if sigRegion is TRUE, a third dataframe is included
    containing the significance region boundaries for each effect as primary,
    moderated by the other.
}
\description{
Calculates Johnson-Neyman significance regions and MOM signficance
  values (raw and Bonferroni-Holm adjusted) for SAOM 2-way
  interactions.
}
