% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoringFuncs.R
\name{PSQIScore}
\alias{PSQIScore}
\title{Calculates PSQI subscales and total scores}
\usage{
PSQIScore(inputTbl, wv = "")
}
\arguments{
\item{inputTbl}{A tbl_df with SSID, then all raw PSQI numeric items in order.\cr
Hence it is N x 19 (SSID + 18 PSQI items).\cr
inputTbl cols 2 & 4 must be of type c('hms', 'difftime' )\cr
inputTbl cols 3 and 5:19 must include either 'double' or 'numeric'
   in their class.\cr
inputTbl col 1 (the ID column), the PSQI ID is normally some numeric
but can be anything.\cr
The 18 input items (ignoring SSID, the first column) must be
included in this order, though they can be named anything:\cr
- psq1\cr
- psq2\cr
- psq3\cr
- psq4\cr
- psq5a to psq5j (10 items)\cr
- psq6\cr
- psq7\cr
- psq8\cr
- psq9\cr}

\item{wv}{Character string with the wave number (defaults to empty string).
This string is appended (with a period) to the component names, which
are all of the form >> pscN << (N = 1, 2, ...7, Tot). So if
This string is nonempty, these names will all be of the form
pscN.v(wv), where v() means "the value of".}
}
\value{
A tbl_df of SSID, PSQI scored items (7), and a total, hence N x 9. \cr
        All columns are converted to 'numeric'.
}
\description{
Calculates PSQI subscales and total scores
}
