% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statFuncs.R
\name{pValT}
\alias{pValT}
\title{Calculates t, 2-tailed p value, pAlpha CI}
\usage{
pValT(pDiff, sE, pAlpha = 0.95, pTails = 2)
}
\arguments{
\item{pDiff}{The difference between the two raw scores of interest, or
a parameter estimate.}

\item{sE}{The applicable standard error for the difference or estimate.}

\item{pTails}{Number of tails (default 2). Any value other than 2 produces
a 1-tailed result.}
}
\value{
A vector with three values: the t-ratio, the two-tailed p-value, and
    the pAlpha confidence limits.
}
\description{
Calculates t, 2-tailed p value, pAlpha CI
}
