% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFuncs.R
\name{pvT}
\alias{pvT}
\title{Calculates P-values}
\usage{
pvT(pDiff, sE, pTails)
}
\arguments{
\item{pDiff}{Number: Effect size}

\item{sE}{Number: standard error}

\item{pTails}{Number: number of tails (1 or 2)}
}
\value{
The p-value (number or vector)
}
\description{
Calculater 1 or 2-tailed pvalues.
}
\details{
A version of JMLUtils::pValT (pvT) which does not output
  anything, just calculates 1 or 2-tailed p-vals.
}
