% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email.R
\name{sendEmail}
\alias{sendEmail}
\title{Send an email (for now, just from John's Gmail account)}
\usage{
sendEmail(subj, msg)
}
\arguments{
\item{subj}{A string with the desired subject line}

\item{pw}{John's secret password}
}
\value{
Nothing. It just sends the Email
}
\description{
Requires running 'setupEmail' first. This function is called after some
other code runs, or whenever you like.
}
