% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFuncs.R
\name{setAllMissing}
\alias{setAllMissing}
\title{Converts missing values to NA, including 88 or 99 (if in network ID vars)}
\usage{
setAllMissing(pInTB)
}
\arguments{
\item{pInTB}{The input tibble from the output of 'inputSPSS'}
}
\value{
The same tibble as was input, except that NaNs (in any vars)
   and 88's or 99's (in network ID variables only, if there are any)
   are all replaced with 'NA'
}
\description{
Converts missing values to NA, including 88 or 99 (if in network ID vars)
}
\details{
This function works for survey and/or network data tables.
}
