% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email.R
\name{setupEmail}
\alias{setupEmail}
\title{Setup auto-sending an Email after some other code runs
(only works for John's Gmail account, for now. Maybe generalized
someday).}
\usage{
setupEmail()
}
\value{
A list containing [[1]] the text message body, [[2]] credential key
}
\description{
Setup auto-sending an Email after some other code runs
(only works for John's Gmail account, for now. Maybe generalized
someday).
}
