% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statFuncs.R
\name{TriadCensus}
\alias{TriadCensus}
\title{Holland and Leinhardt Triad Census; see ?sna::triad.census.}
\usage{
TriadCensus(i, data, sims, wave, groupName, varName, levls = 1:16)
}
\description{
Holland and Leinhardt Triad Census; see ?sna::triad.census.
}
