% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixRates.R
\name{unfixRates}
\alias{unfixRates}
\title{Frees rate parameters (e.g. that were fixed by 'fixRates')}
\usage{
unfixRates(efObj, efxNums = c(0))
}
\arguments{
\item{efObj}{An RSiena 'effects' object.}

\item{efxNums}{A vector of effect numbers of the rates to be fixed. If
this vector consists only of a 1 element vector with 0 in it, or is
omitted (in which case, c(0) is the default), all rate parameters
in the effects object are fixed, meaning they are not estimated,
but remain at their starting values.}
}
\value{
The modified effects object, with specified (or all ) rate
  parameters freed.
}
\description{
Frees rate parameters (e.g. that were fixed by 'fixRates')
}
