% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFuncs.R
\name{wksp.size}
\alias{wksp.size}
\title{Return the size of the workspace environment currently loaded}
\usage{
wksp.size()
}
\value{
A number (size of current wksp in bytes).
   Wide Format' means that each chooser has only 1 row per wave, and the
   relationship choices directed to other house members are all contained
   in that one row. The data.table is basically an image of the SPSS file
   (.sav) that it reads.
}
\description{
This is usually the first function run when creating a network analysis.
}
