% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFuncs.R
\name{getglmmTMBFit}
\alias{getglmmTMBFit}
\title{Pulls glmmTMB model results with t and p-vals}
\usage{
getglmmTMBFit(fitObj, fxNames, alpha = 0.05, tails = 2)
}
\arguments{
\item{fitObj}{An RSiena 'sienaFit' object (output from siena07)}

\item{fxNames}{A character vector the same length as the total number
of fixed effects in the model. For the future: These names are usually
named elements of the parameter vector. However, you'd usually want
custom row names in any published table.}

\item{alpha}{A number (Default=.05) to use for creating confidence
intervals around the parameter}

\item{tails}{Numeric: default is 2; enter 1 if 1-tailed pval preferred.}
}
\value{
A tibble with 8 columns (if tails = 1) or 10 columns (if tails
  =2; only in this case are alpha-level CIs calculated and output, as
  well as alpha itself).
}
\description{
Extracts effect numbers, effect types, effect names,
  thetas (estimated parameters) and standard errors from a
  glmmTMB object, then calculates t-statistics, one or two-tailed
  p-values, and (if tails = 2) an alpha-level confidence interval
  for each effect. The alpha level is also ouput, for later reference.
}
